submissions/task-4.1/submission-4.1/agent/requirements.txt
openai>=1.40.0
PyYAML>=6.0.0
hjson>=3.1.0
